+++ app/(tabs)/index.tsx
@@ -1,51 +1,18 @@
 import { useState } from 'react';
 import { Image } from 'expo-image';
-import { Platform, StyleSheet, Pressable, Alert, View } from 'react-native';
-import { useRouter } from 'expo-router';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-
+import { Platform, StyleSheet, View } from 'react-native';
 import { HelloWave } from '@/components/hello-wave';
 import ParallaxScrollView from '@/components/parallax-scroll-view';
 import { ThemedText } from '@/components/themed-text';
 import { ThemedView } from '@/components/themed-view';
 import { Link } from 'expo-router';
-import { useColorScheme } from '@/hooks/use-color-scheme';
 import { useAccount } from '@spezivibe/account';
-import { AccountButton } from '@/app/(account)/_components/account-button';
-import { AccountSheet } from '@/app/(account)/_components/account-sheet';
-import { ONBOARDING_COMPLETED_KEY } from '@/lib/constants';
+import { AccountButton } from '@/components/account/account-button';
+import { AccountSheet } from '@/components/account/account-sheet';
 
 export default function HomeScreen() {
-  const router = useRouter();
-  const colorScheme = useColorScheme();
-  const isDark = colorScheme === 'dark';
   const { signedIn } = useAccount();
   const [showAccountSheet, setShowAccountSheet] = useState(false);
-
-  const handleViewOnboarding = async () => {
-    Alert.alert(
-      'View Onboarding',
-      'This will clear the onboarding state and show the onboarding flow.',
-      [
-        { text: 'Cancel', style: 'cancel' },
-        {
-          text: 'View Onboarding',
-          onPress: async () => {
-            try {
-              await AsyncStorage.removeItem(ONBOARDING_COMPLETED_KEY);
-
-              // Add a small delay to ensure the root layout picks up the change
-              setTimeout(() => {
-                router.push('/(onboarding)/welcome');
-              }, 150);
-            } catch (error) {
-              // Silent error
-            }
-          },
-        },
-      ]
-    );
-  };
 
   return (
     <>
@@ -111,24 +78,6 @@ export default function HomeScreen() {
           {`Tap the Explore tab to learn more about what's included in this starter app.`}
         </ThemedText>
       </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 3: View Onboarding</ThemedText>
-        <ThemedText>
-          Want to see the onboarding flow? Tap the button below to clear the onboarding state
-          and view it again.
-        </ThemedText>
-        <Pressable
-          style={({ pressed }) => [
-            styles.button,
-            { backgroundColor: isDark ? '#B83A4B' : '#8C1515', opacity: pressed ? 0.8 : 1, marginTop: 12 },
-          ]}
-          onPress={handleViewOnboarding}>
-          <ThemedText style={[styles.buttonText, { color: isDark ? '#000' : '#fff' }]}>
-            View Onboarding Flow
-          </ThemedText>
-        </Pressable>
-      </ThemedView>
-
     </ParallaxScrollView>
 
       <AccountSheet
